22/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "receipt-ocr-agent"
version = "0.1.0"
description = "領収書OCRエージェント"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.31.0",
    "langchain-core>=0.1.17",
    "langchain-anthropic>=0.1.9",
    "langgraph>=0.0.20",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "pandas>=2.0.0",
    "pydantic>=2.5.0",
]

[tool.uv]
dev-dependencies = [
    "pandas-stubs>=2.2.3.250308",
    "black>=23.1.0",
    "isort>=5.12.0",
    "mypy>=1.1.1",
]

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
